Make header shorter
