+++ src/styles/common/grid.scss
@@ -1,7 +1,7 @@
 @use '../mixins';
 
 :root {
-  --header-height: 5rem;
+  --header-height: 4rem;
   --navigation-height: 3em;
   --footer-height: 3em;
   --main-margin: 2em;
@@ -10,7 +10,7 @@
 
 @include mixins.mq(sp) {
   :root {
-    --header-height: 3.5rem;
+    --header-height: 3rem;
     --navigation-height: 0px;
     --footer-height: 3em;
     --main-margin: 8px;
